implemented rating API for client

+++ src/main/java/de/crow08/musicstreamserver/rating/Rating.java
@@ -34,7 +34,7 @@ public class Rating {
     this.id = id;
   }
 
-  public long getRatingValue() {
+  public short getRatingValue() {
     return ratingValue;
   }
